Refactor modules into package
