+++ tests/test_small_sections.py
@@ -1,6 +1,7 @@
 import importlib.util
 import sys
 import random
+import types
 from pathlib import Path
 
 import pytest
@@ -8,15 +9,27 @@ import pytest
 BASE = Path(__file__).resolve().parent.parent
 
 def _load(name: str, file: str):
+    pkg = name.split(".")[0]
+    if pkg not in sys.modules:
+        package = types.ModuleType(pkg)
+        package.__path__ = [str(BASE / "tripd")]
+        sys.modules[pkg] = package
     spec = importlib.util.spec_from_file_location(name, BASE / file)
     module = importlib.util.module_from_spec(spec)
     sys.modules[name] = module
     spec.loader.exec_module(module)
+    parent = sys.modules[pkg]
+    parts = name.split(".")
+    if len(parts) > 1:
+        setattr(parent, parts[1], module)
+        if file.endswith("__init__.py"):
+            for sym in getattr(module, "__all__", []):
+                setattr(parent, sym, getattr(module, sym))
     return module
 
-memory = _load("tripd_pkg.tripd_memory", "tripd_memory.py")
-expansion = _load("tripd_pkg.tripd_expansion", "tripd_expansion.py")
-tripd = _load("tripd_pkg.tripd", "tripd.py")
+memory = _load("tripd_pkg.tripd_memory", "tripd/tripd_memory.py")
+expansion = _load("tripd_pkg.tripd_expansion", "tripd/tripd_expansion.py")
+tripd = _load("tripd_pkg.tripd", "tripd/__init__.py")
 TripDModel = tripd.TripDModel
