+++ tripd/tripd_expansion.py
