+++ tripd/tripdictionary.md
